Created standardized cargo plate format
Also created generic driver for these cargo plates

Fixed old plates so that they were updated. New
ones can be found in the "used-cargo-plates"
folder. Generic driver is currently just floating
in the GUI folder, and not implemented in the
main code yet. Will integrate it soon (hopefully
tomorrow). But in the meantime, it seems able to
import plates from a standard excel file.